+++ helper/src/command/docker.rs
@@ -3,10 +3,9 @@ use std::process::{exit, Command, Stdio};
 use anyhow::{Context, Result};
 use cargo_metadata::camino::Utf8PathBuf;
 
-use crate::{
-    program::{get_program_build_args, get_rust_compiler_flags},
-    BuildArgs,
-};
+use crate::BuildArgs;
+
+use super::utils::{get_program_build_args, get_rust_compiler_flags};
 
 /// Uses SP1_DOCKER_IMAGE environment variable if set, otherwise constructs the image to use based
 /// on the provided tag.
@@ -60,7 +59,13 @@ pub(crate) fn create_docker_command(
         "docker"
     );
 
-    // Add docker-specific arguments.
+    // When executing the Docker command:
+    // 1. Set the target directory to a subdirectory of the program's target directory to avoid build
+    // conflicts with the parent process. Source: https://github.com/rust-lang/cargo/issues/6412
+    // 2. Set the rustup toolchain to succinct.
+    // 3. Set the encoded rust flags.
+    // Note: In Docker, you can't use the .env command to set environment variables, you have to use
+    // the -e flag.
     let mut docker_args = vec![
         "run".to_string(),
         "--platform".to_string(),
